% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-jartic_provider.R
\docType{data}
\name{jartic_provider}
\alias{jartic_provider}
\title{Jartic data provider}
\format{
A tibble with 51 rows 2 variables:
\itemize{
  \item{name: prefecture}
  \item{name_en: roman}
}
}
\usage{
jartic_provider
}
\description{
data provider list
}
\examples{
jartic_provider
}
\keyword{datasets}
