% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/location.R
\name{jartic_type_b_loc_tiny}
\alias{jartic_type_b_loc_tiny}
\title{Create location table for type B}
\usage{
jartic_type_b_loc_tiny(data)
}
\arguments{
\item{data}{type B data}
}
\description{
Create location table for type B
}
