% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_jartic_trafifc.R
\name{read_jartic_traffic}
\alias{read_jartic_traffic}
\title{Read jartic traffic file}
\usage{
read_jartic_traffic(path)
}
\arguments{
\item{path}{path to raw data (csv format)}
}
\value{
data.frame (data.table)
}
\description{
read type B file
}
